+++ src/main/java/naeilmolae/global/infrastructure/kakao/KakaoClient.java
@@ -1,0 +1,4 @@
+package naeilmolae.global.infrastructure.kakao;
+
+public class KakaoClient {
+}
