+++ src/main/java/naeilmolae/global/infrastructure/kakao/dto/KakaoResponse.java
@@ -1,4 +1,4 @@
-package naeilmolae.domain.member.dto.client;
+package naeilmolae.global.infrastructure.kakao.dto;
 
 import lombok.Getter;
